requirements.txt
pandas
sentence-transformers
numpy
tornado
requests
groq
python-dotenv